hima: Set PROC_BTWRITE_TIMER_TIMEOUT_MS to 0
Set PROC_BTWRITE_TIMER_TIMEOUT_MS to 0
to avoid upio start and stop timer when LMP
is enabled and BT_WAKE is asserted and
de-asserted. This time is useless and would
set btwrite to 0 by mistake even when host
is sending data to UART, and this would
de-assert BT_WAKE by mistake.

Change-Id: Icb8f638a90d24016e770f60cb649d79af5e5104d